Added realign and unwarp interface to spm
I've added the realign and unwarp interface to spm/preprocess.py. I tried it on my data and in works. I wanted to also implement the field map toolbox of spm, but I didn't get around to doing it. There might just be an implementation of a voxel displacement map calculation which can just be fed into the realign and unwarp interface.